Parser: Fix register value not handled correctly with error "InvalidRegisterNumber"
Erroneous downcast from std::int32_t to std::int8_t made the resulting
error message incorrect.
This was because I forgot to change all occurrences of std::int8_t -- related
to register parsing information -- to std::int32_t.

+++ libmomiji/src/Parser/Instructions.cpp
@@ -32,7 +32,7 @@ namespace momiji::details
                 metadata.error  = std::move(error);
             };
 
-            std::int8_t regval = momiji::extractRegister(operand);
+            std::int32_t regval = momiji::extractRegister(operand);
 
             if (checkRegRange(regval))
             {
